requirements.txt
sounddevice
numpy
faster-whisper
soundfile
librosa
speechbrain
torch
websockets
silero-vad
noisereduce
fastapi
pydantic
uvicorn
PyJWT==2.10.1
